workaround desktop nav not being clickable
